+++ tests/core/java-testsuite-runner-core/CHANGELOG.rst
@@ -17,6 +17,7 @@ Added
 - Add check for ``LLMJVM_IMPL_setApplicationTime``.
 - Test Thread.sleep() with max number of milliseconds (Long.MAX_VALUE) does not cause an infinite loop in the MicroEJ Core due to time conversion overflow.
 - Add check for ``LLMJVM_IMPL_getCurrentTime``, which should not go back in time.
+- Add check for C malloc heap configuration which can be used when parsing float.
 
 Changed
 ~~~~~~~